refactor: move characterData.ts to utils
