+++ src/utils/characterData.ts
